+++ CHANGELOG.md
@@ -4,6 +4,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bump runtimeVersion to `2025.06.20`.
  * Upgraded stable Flutter analysis SDK to `3.32.4`.
+ * Note: started to export `/api/packages/<package>/[likes|options|publisher|score]` endpoints.
 
 ## `20250619t085100-all`
  * Bump runtimeVersion to `2025.06.03`.